+++ cli/.js/src/main/scala/aqua/run/RunOpts.scala
@@ -3,6 +3,7 @@ package aqua.run
 import aqua.*
 import aqua.ArgOpts.checkDataGetServices
 import aqua.builder.{ArgumentGetter, Service}
+import aqua.io.{AquaPath, RelativePath}
 import aqua.model.transform.TransformConfig
 import aqua.model.{LiteralModel, ValueModel, VarModel}
 import aqua.parser.expr.func.CallArrowExpr
